refactor(shared): split dev_server.py (997 LOC) into 10-file package (Campaign B B4)
Iterate Campaign B sub-iterate B4 — campaign
`2026-05-25-bloat-cleanup-B-shipwright`. Splits the 997-LOC
`shared/scripts/dev_server.py` monolith into a focused package
where every submodule is well under the 300-LOC source budget.

Package layout (was 997 LOC -> all submodules <= 300 LOC each):

  shared/scripts/dev_server.py            29 LOC  re-export shim for uv run
  shared/scripts/dev_server/
    __init__.py                          187 LOC  re-exports public surface
    __main__.py                           16 LOC  `python -m` entry
    cli.py                                94 LOC  argparse + --services-json
    health.py                            124 LOC  TCP + HTTP probes
    spawn.py                             179 LOC  process spawn/kill/rollback
    profile_config.py                    272 LOC  profile load + normalize
    validation.py                        128 LOC  schema + topo sort
    state.py                              86 LOC  v1-read/v2-atomic IO
    multiservice.py                      267 LOC  cmd_start/stop/status
    _proxies.py                           68 LOC  package-surface dispatch

Why a shim + package side by side:
Five skill .md files invoke `uv run shared/scripts/dev_server.py` by
absolute path. Deleting the file would break every one of those
callers. Python's import system prefers a package directory over a
same-named .py file on the same sys.path entry; `import dev_server`
from the test conftest resolves to the package, while
`uv run shared/scripts/dev_server.py` executes the shim (file-path
execution, not import). The new
test_b4_split_import_resolves_to_package_not_shim asserts this
precedence empirically.

Why _proxies.py:
The existing test suite heavily monkeypatches the package surface
(@patch("dev_server._kill_one"),
monkeypatch.setattr("dev_server.subprocess.Popen", ...)). Direct
`from .health import _is_port_in_use_for_host` in multiservice.py
would hold a stale reference and miss the patch. _proxies.py holds
tiny dispatch stubs that look names up on sys.modules[__package__]
at call time, so test monkeypatches propagate into submodule code
paths without rewriting any of the 80 existing dev_server tests.

B3 lesson respected:
No submodule does module-level `from lib.cmd_resolver import ...`.
resolve_executable is bound lazily on the package and the
lib.cmd_resolver import happens at first call, never during pytest
collection.

Verification:
- F0.5 surface_verification (surface=cli, tests_run=80, exit_code=0).
- 80 dev_server tests pass (78 pre-existing + 2 new B4 parity tests in
  a dedicated 78-LOC file), 1 pre-existing intentional skip.
- `uv run shared/scripts/dev_server.py --help` identical pre/post-split.
- `uv run python -m shared.scripts.dev_server --help` works.
- `uv run shared/scripts/dev_server.py stop --cwd <tmp>` returns
  JSON identical to pre-split shape.
- External LLM iterate review (OpenRouter; openai + gemini): 12
  findings, dispositions tracked in ADR-062.
- shared/tests/ full suite: 2435 passed, 2 pre-existing failures
  on main unrelated to B4.
- plugins/shipwright-compliance/tests/: 553 passed (B3 pollution
  hazard confirmed absent).

Cleanup-invariant:
shipwright_bloat_baseline.json entry for shared/scripts/dev_server.py
REMOVED in this same atomic commit per rule (b) — the path now exists
at 29 LOC (well under the 300-LOC budget) and the bulk lives in the
new package directory. New parity tests live in a separate file to
avoid ratcheting the multiservice test file's baseline.

ADR: .shipwright/planning/adr/062-b4-dev-server-split.md
