+++ CHANGELOG-unreleased.d/Changed/iterate-2026-05-25-bloat-B-dev-server_001.md
@@ -1,0 +1,1 @@
+- (refactor) Split `shared/scripts/dev_server.py` from 997 LOC into a 10-file package (`shared/scripts/dev_server/`): `__init__.py` (re-exports the pre-split public surface + lazy-bound `resolve_executable`), `__main__.py` (`python -m shared.scripts.dev_server` entry), `cli.py` (argparse + `--services-json` handling), `health.py` (TCP port + HTTP readiness probes), `spawn.py` (process spawn / kill / rollback), `profile_config.py` (profile JSON loading + `${VAR:-default}` expansion + service-entry normalization + legacy `_get_config`), `validation.py` (service-list schema validation + topo sort + `_pick_primary`), `state.py` (`shipwright_dev_server.json` v1-read-compat + v2 atomic write IO), `multiservice.py` (top-level `cmd_start` / `cmd_start_with_services` / `cmd_stop` / `cmd_status` orchestration), `_proxies.py` (internal proxy stubs that dispatch through the package surface so test monkeypatches at `dev_server.X` propagate into submodule code paths). Each submodule <= 300 LOC. `shared/scripts/dev_server.py` reduced to a 29-LOC shim so `uv run shared/scripts/dev_server.py <args>` keeps working unchanged (Python's import system prefers the package directory over the same-named file; empirically verified by a new parity test). Public import surface preserved — every name reachable via `dev_server.X` pre-split is re-exported. `resolve_executable` is bound lazily at the package level (B3 lesson — no module-level `from lib.cmd_resolver import` in any submodule pytest may collect). `shipwright_bloat_baseline.json` entry removed per cleanup invariant rule (b). Two new B4 parity tests (`test_b4_split_public_surface_present`, `test_b4_split_import_resolves_to_package_not_shim`) guard the surface contract empirically. Campaign B sub-iterate B4.
